feat: add non-apt-tied behavior annotation and configuration to core
Fixes #1012
