+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/config/runtime/DefaultConfigurationService.java
@@ -3,52 +3,43 @@ package io.javaoperatorsdk.operator.config.runtime;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.javaoperatorsdk.operator.api.config.BaseConfigurationService;
 import io.javaoperatorsdk.operator.api.config.ControllerConfiguration;
-import io.javaoperatorsdk.operator.api.config.Utils;
 import io.javaoperatorsdk.operator.api.reconciler.Reconciler;
 
 public class DefaultConfigurationService extends BaseConfigurationService {
 
   private static final DefaultConfigurationService instance = new DefaultConfigurationService();
+  private boolean createIfNeeded = super.createIfNeeded();
 
   private DefaultConfigurationService() {
-    super(Utils.loadFromProperties());
+    super();
   }
 
   public static DefaultConfigurationService instance() {
     return instance;
   }
 
-  @Override
-  public <R extends HasMetadata> ControllerConfiguration<R> getConfigurationFor(
-      Reconciler<R> reconciler) {
-    return getConfigurationFor(reconciler, true);
-  }
-
   <R extends HasMetadata> ControllerConfiguration<R> getConfigurationFor(
       Reconciler<R> reconciler, boolean createIfNeeded) {
-    var config = super.getConfigurationFor(reconciler);
-    if (config == null) {
-      if (createIfNeeded) {
-        // create the configuration on demand and register it
-        config = new AnnotationControllerConfiguration<>(reconciler);
-        register(config);
-        getLogger().info(
-            "Created configuration for reconciler {} with name {}",
-            reconciler.getClass().getName(),
-            config.getName());
-      }
-    } else {
-      // check that we don't have a reconciler name collision
-      final var newControllerClassName = reconciler.getClass().getCanonicalName();
-      if (!config.getAssociatedReconcilerClassName().equals(newControllerClassName)) {
-        throwExceptionOnNameCollision(newControllerClassName, config);
-      }
+    final var previous = createIfNeeded();
+    setCreateIfNeeded(createIfNeeded);
+    try {
+      return super.getConfigurationFor(reconciler);
+    } finally {
+      setCreateIfNeeded(previous);
     }
-    return config;
   }
 
   @Override
-  public boolean checkCRDAndValidateLocalModel() {
-    return Utils.shouldCheckCRDAndValidateLocalModel();
+  protected boolean createIfNeeded() {
+    return createIfNeeded;
+  }
+
+  public void setCreateIfNeeded(boolean createIfNeeded) {
+    this.createIfNeeded = createIfNeeded;
+  }
+
+  @Override
+  protected <R extends HasMetadata> ControllerConfiguration<R> configFor(Reconciler<R> reconciler) {
+    return new AnnotationControllerConfiguration<>(reconciler);
   }
 }
